Migrate more users endpoints to new pattern
