+++ server/service/transport_users.go
@@ -10,7 +10,7 @@ import (
 
 func decodeCreateUserRequest(ctx context.Context, r *http.Request) (interface{}, error) {
 	var req createUserRequest
-	if err := json.NewDecoder(r.Body).Decode(&req.payload); err != nil {
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
 		return nil, err
 	}
